requirements.txt
pypdfium2
pdfplumber
Pillow
torch
transformers
